chore(network): rename streamed_data_protocol to streamed_data
