+++ crates/papyrus_network/src/streamed_data/flow_test.rs
@@ -11,7 +11,7 @@ use libp2p::{PeerId, Swarm};
 use super::behaviour::{Behaviour, Event};
 use super::{InboundSessionId, OutboundSessionId, SessionId};
 use crate::messages::protobuf;
-use crate::streamed_data_protocol::Config;
+use crate::streamed_data::Config;
 use crate::test_utils::{create_fully_connected_swarms_stream, StreamHashMap};
 
 const NUM_PEERS: usize = 3;
